Initial code
This takes accel, mag and light sensor as input
to animate two led strips. Right now they are
animated identically, but it is prepared to make
them different.

It uses the accel to only change lights when it moves
and chooses the color based on the mag sensor.

It also uses the light sensor to play an animatino
when it moves from a dark to a light environment and
vice versa.